flask predictor model/requirements.txt
pandas
scipy
sklearn
fuzzywuzzy
flask
imdbpy